+++ Lib.Test/TSCompilerTests.cs
@@ -158,12 +158,12 @@ namespace Lib.Test
         {
             _bbdir = PathUtils.Join(PathUtils.Normalize(Environment.CurrentDirectory), ".bbcore");
             _tools = new ToolsDir.ToolsDir(PathUtils.Join(_bbdir, "tools"));
-            _tools.SetTypeScriptVersion("2.7.2");
+            _tools.SetTypeScriptVersion(TSProject.DefaultTypeScriptVersion);
             _compilerPool = new CompilerPool(_tools);
         }
 
         [Fact]
-        public void LatestTypeScriptVersionDidntChanged()
+        public void DefaultTypeScriptVersionDidntChanged()
         {
             Assert.Equal("2.7.2", _tools.TypeScriptVersion);
         }